requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
pydantic[email]==2.4.2
pydantic-settings==2.0.3
httpx==0.25.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
tavily-python==0.1.9

# Dev dependencies
pytest==7.4.2
black==23.9.1
flake8==6.1.0